+++ sos21-database/migrations/20210615083430_add_project_query_file_sharing.sql
@@ -1,0 +1,8 @@
+ALTER TYPE file_sharing_scope ADD VALUE 'project_query';
+COMMIT; -- we need to commit current transaction before we use 'project_query' enum value
+
+ALTER TABLE file_sharings
+    ADD COLUMN project_query jsonb,
+    ADD CONSTRAINT file_sharings_scope_project_query CHECK (
+        (scope = 'project_query') = (project_query IS NOT NULL)
+    );