+++ sos21-database/src/model/file_sharing.rs
@@ -6,6 +6,7 @@ use uuid::Uuid;
 #[sqlx(rename_all = "snake_case")]
 pub enum FileSharingScope {
     Project,
+    ProjectQuery,
     FormAnswer,
     RegistrationFormAnswer,
     Committee,
@@ -22,6 +23,7 @@ pub struct FileSharing {
     pub expires_at: Option<DateTime<Utc>>,
     pub scope: FileSharingScope,
     pub project_id: Option<Uuid>,
+    pub project_query: Option<serde_json::Value>,
     pub form_answer_project_id: Option<Uuid>,
     pub form_answer_form_id: Option<Uuid>,
     pub registration_form_answer_project_id: Option<Uuid>,