+++ sos21-domain/src/model/file_sharing.rs
@@ -153,8 +153,8 @@ impl FileSharing {
         Ok(())
     }
 
-    pub fn scope(&self) -> FileSharingScope {
-        self.0.scope
+    pub fn scope(&self) -> &FileSharingScope {
+        &self.0.scope
     }
 
     /// Equivalent to `sharing.state_at(DateTime::now())`
@@ -406,6 +406,41 @@ mod tests {
         let user = test_model::new_operator_user();
         let sharing = FileSharing::new(test_model::new_file_id(), FileSharingScope::Committee);
         assert!(sharing.to_witness_with_user(&user).is_ok());
+    }
+
+    #[test]
+    fn test_witness_with_project_query_ok() {
+        use crate::model::project::{ProjectAttributes, ProjectCategory};
+        use crate::model::project_query::{ProjectQuery, ProjectQueryConjunction};
+
+        let query = ProjectQuery::from_conjunctions(vec![ProjectQueryConjunction {
+            category: Some(ProjectCategory::General),
+            attributes: ProjectAttributes::from_attributes(vec![]).unwrap(),
+        }])
+        .unwrap();
+        let scope = FileSharingScope::ProjectQuery(query);
+        let sharing = FileSharing::new(test_model::new_file_id(), scope);
+        let project = test_model::new_general_project(test_model::new_user_id());
+        assert!(sharing.to_witness_with_project(&project).is_ok());
+    }
+
+    #[test]
+    fn test_witness_with_project_query_out_of_scope() {
+        use crate::model::project_query::ProjectQuery;
+
+        use super::ToWitnessErrorKind;
+
+        let query = ProjectQuery::from_conjunctions(vec![]).unwrap();
+        let scope = FileSharingScope::ProjectQuery(query);
+        let sharing = FileSharing::new(test_model::new_file_id(), scope);
+        let project = test_model::new_general_project(test_model::new_user_id());
+        assert!(matches!(
+            sharing
+                .to_witness_with_project(&project)
+                .unwrap_err()
+                .kind(),
+            ToWitnessErrorKind::OutOfScope
+        ));
     }
 
     #[test]