+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Added
+
+- `project_query` file sharing ([#104](https://github.com/sohosai/sos21-backend/pull/104))
+
 ### Changed
 ### Deprecated
 ### Removed